fix(review-queue): split dissent verdict parser
